Renamed the rules.py to user_sync.py and added to the module user_sync/engine
