+++ user_sync/app.py
@@ -40,7 +40,7 @@ import user_sync.encryption
 import user_sync.helper
 import user_sync.lockfile
 import user_sync.resource
-import user_sync.rules
+import user_sync.engine.user_sync
 
 import user_sync.connector.umapi
 from user_sync.post_sync.manager import PostSyncManager
@@ -359,7 +359,7 @@ def begin_work(config_loader):
     for groups in six.itervalues(directory_groups):
         for group in groups:
             umapi_name = group.umapi_name
-            if umapi_name != user_sync.rules.PRIMARY_UMAPI_NAME:
+            if umapi_name != user_sync.engine.user_sync.PRIMARY_UMAPI_NAME:
                 referenced_umapi_names.add(umapi_name)
     referenced_umapi_names.difference_update(six.iterkeys(secondary_umapi_configs))
     if len(referenced_umapi_names) > 0:
@@ -408,9 +408,9 @@ def begin_work(config_loader):
         umapi_secondary_conector = user_sync.connector.umapi.UmapiConnector(".secondary.%s" % secondary_umapi_name,
                                                                             secondary_config)
         umapi_other_connectors[secondary_umapi_name] = umapi_secondary_conector
-    umapi_connectors = user_sync.rules.UmapiConnectors(umapi_primary_connector, umapi_other_connectors)
+    umapi_connectors = user_sync.engine.user_sync.UmapiConnectors(umapi_primary_connector, umapi_other_connectors)
 
-    rule_processor = user_sync.rules.RuleProcessor(rule_config)
+    rule_processor = user_sync.engine.user_sync.RuleProcessor(rule_config)
     if len(directory_groups) == 0 and rule_processor.will_process_groups():
         logger.warning('No group mapping specified in configuration but --process-groups requested on command line')
     rule_processor.run(directory_groups, directory_connector, umapi_connectors)
