+++ user_sync/post_sync/connectors/sign_sync/__init__.py
@@ -2,7 +2,7 @@ import logging
 from collections import defaultdict
 from user_sync.post_sync import PostSyncConnector
 from user_sync.config import DictConfig, OptionsBuilder
-from user_sync.rules import AdobeGroup
+from user_sync.engine.user_sync import AdobeGroup
 from .client import SignClient
 from user_sync.error import AssertionException
